Advanced Programming Methods (MAP)/Labs/Lab3-MaziluPaul/src/Domain/Pacient.java
package Domain;

import java.util.Objects;

public class Pacient extends Entity{
    private static final long serialVersionUID = 1000L;
    private String nume;
    private String prenume;
    private int varsta;

    public Pacient(int id, String nume, String prenume, int varsta) {
        super(id);
        this.nume = nume;
        this.prenume = prenume;
        this.varsta = varsta;
    }

    public String getNume() {
        return nume;
    }

    public void setNume(String nume) {
        this.nume = nume;
    }

    public String getPrenume() {
        return prenume;
    }

    public void setPrenume(String prenume) {
        this.prenume = prenume;
    }

    public int getVarsta() {
        return varsta;
    }

    public void setVarsta(int varsta) {
        this.varsta = varsta;
    }

    @Override
    public String toString() {
        return "Pacient{" +
                "id=" + id +
                ", nume='" + nume + '\'' +
                ", prenume='" + prenume + '\'' +
                ", varsta=" + varsta +
                '}';
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Pacient pacient = (Pacient) o;
        return this.getId() == pacient.getId() && Objects.equals(this.getNume(), pacient.getNume()) && Objects.equals(this.getPrenume(), pacient.getPrenume()) && this.varsta == pacient.getVarsta();
    }
}
